fix(db): widen log_index columns from integer to bigint
archive_confirmation.log_index and ingestion_dlq{,_resolved}.archive_log_index
were integer (signed 32-bit, max 2,147,483,647). The Compound Governor backfill
hit an event with logIndex = 4294967295 (0xFFFFFFFF, uint32 max), which overflows
a signed integer column and halts the backfill run.

+++ libs/db/migrations/0002_core_domain.ts
@@ -215,7 +215,7 @@ export async function up(db: Kysely<unknown>): Promise<void> {
     .addColumn('block_number', 'bigint', (col) => col.notNull())
     .addColumn('block_hash', 'text', (col) => col.notNull())
     .addColumn('tx_hash', 'text', (col) => col.notNull())
-    .addColumn('log_index', 'integer', (col) => col.notNull())
+    .addColumn('log_index', 'bigint', (col) => col.notNull())
     .addColumn('event_type', 'text', (col) => col.notNull())
     .addColumn('received_at', 'timestamptz', (col) => col.notNull())
     .addColumn('confirmation_status', sql`confirmation_status`, (col) => col.notNull())
